+++ packages/components/src/components/spinner/spinner.tsx
@@ -8,9 +8,17 @@ import {
   validateProps,
 } from '../../utils';
 import { getComponentCss } from './spinner-styles';
-import { SPINNER_ARIA_ATTRIBUTES, SPINNER_SIZES, type SpinnerAriaAttribute, type SpinnerSize } from './spinner-utils';
+import {
+  SPINNER_ARIA_ATTRIBUTES,
+  SPINNER_COLORS,
+  SPINNER_SIZES,
+  type SpinnerAriaAttribute,
+  type SpinnerColor,
+  type SpinnerSize,
+} from './spinner-utils';
 
 const propTypes: PropTypes<typeof Spinner> = {
+  color: AllowedTypes.oneOf<SpinnerColor>(SPINNER_COLORS),
   size: AllowedTypes.breakpoint<SpinnerSize>(SPINNER_SIZES),
   aria: AllowedTypes.aria<SpinnerAriaAttribute>(SPINNER_ARIA_ATTRIBUTES),
 };
@@ -22,7 +30,10 @@ const propTypes: PropTypes<typeof Spinner> = {
 export class Spinner {
   @Element() public host!: HTMLElement;
 
-  /** Size of the spinner. */
+  /** Basic color variations. */
+  @Prop() public color?: SpinnerColor = 'primary';
+
+  /** Defines the size of the spinner, aligned with the typographic scale used by components such as p-icon, p-flag, p-text, and p-heading. When set to `inherit`, the size is derived from a custom font-size defined on a parent element, calculated against the global line-height (based on `ex`-unit) to remain visually consistent with other typographic-scale-based components. */
   @Prop() public size?: BreakpointCustomizable<SpinnerSize> = 'small';
 
   /** Add ARIA attributes. */
@@ -34,17 +45,17 @@ export class Spinner {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    attachComponentCss(this.host, getComponentCss, this.size);
+    attachComponentCss(this.host, getComponentCss, this.color, this.size);
 
     return (
-      <span class="root" role="alert" aria-live="assertive" {...parseAndGetAriaAttributes(this.aria)}>
+      <div role="alert" aria-live="assertive" {...parseAndGetAriaAttributes(this.aria)}>
         {/* empty element needed to announce aria-label in screen readers */}
         <span class="sr-only">&nbsp;</span>
         <svg viewBox="-16 -16 32 32" width="100%" height="100%" focusable="false" aria-hidden="true">
-          <circle r="9" />
-          <circle r="9" />
+          <circle r="11" />
+          <circle r="11" />
         </svg>
-      </span>
+      </div>
     );
   }
 }